initial controversy detection
i've based my code primarily on prior art so far, specifically:
https://github.com/jmanuoz/Measuring-controversy-in-Social-Networks-through-NLP

and also a little bit of:
https://github.com/gvrkiran/controversy-detection

By replicating the first solution and making some changes to the
code to update it and make it more robust, i am able to achieve
some simple controversy detection. i have implemented the fasttext
variant (the default) from that project, can give a search term
to the CLI interface, and craft a controversy score that indicates
how controversial tweets that matched that term are. this project
performs data collection and graph construction in R, and then
scores the controversy of the text in python.

i want to make a few improvements, such as adding the BERT variant
from the original project, and see what else there is i can do
to make the CLI more robust for queries.