more advanced arguments
with this commit, more advanced arguments can be supplied to the
command line interface. arguments with spaces, certain special
characters (right now, just apostrophe/single quote and pound sign
[aka hashtag for twitter]) can be included as well. the right files
will be generated, and the code will be able to parse the inputs
as needed. i had to make some changes to not only the r code, but
i also created an additional python file to deal with running the
fasttext executable, so i could more easily manage the input.

i've also included a requirements.txt for my python environment,
and moved my r_requirements.R into the submission folder. i still
have a bit more i want to try to do, and i have a lot of
documentation to add. if time permits i may even try to containerize
the code and make it simpler to run, setting environment variables
for things like number of tweets and app secrets.