containerization and documentation
containerization finally works! that was more painful than expected...
anyway, this is working with podman, which is generally easier for
me to use on my Fedora Linux distribution. luckily, podman should
be a drop-in replacement for docker, so even without the alias
docker=podman, this should work for docker users as well.

this commit includes all the changes required to make the podman
containers build and run (including a script that shows how to do
this). i now provide the search term as an environment variable
to the container, allowing the dockerfile to parse that environment
variable and pass it in to the required controversy detection
scripts. i lowered the number of tweets to collect to make the
process shorter. the longest part by far is the installation of
the r packages, but i tried to speed that up by using multiple
cores. either way, this code can now be run locally with the right
setup, or in a container for a platform agnostic solution.

i also added the required documentation. i still need to make my
video and include that in the documentation, so there is still one
more commit to come. but this documentation should satisfy all of
the other requirements for the documentation aspect of the rubric,
at least as far as i can tell. this documentation includes motivation,
description of the code, and setup and usage. i wanted to commit
all of this working code before shooting my video.