adding presentation video to readme; final submission
i have made my presentation video which shows running the code and
describes the process for getting everything to work. i show some
sample use cases, such as searching for "#XenobladeChronicles3"
which is a new video game that has a niche "popularity".

i elaborated a bit more in my readme, fixing some mistakes and
adding a section about possible bugs that i encountered during
the process, though hopefully the code is now robust to many of
those bugs...hopefully.

finally, in the vein of academic integrity i added the source of
some of the files i extracted specifically from my referenced
sources to the top of those files. i think i hit on all of the files,
but it's possible i missed one...i call out in the readme that much
of the code comes from prior art, which is referenced explicitly.
with this, i think my project is pretty much done! i'm interested
to do more with this in the future though!!