privacy disclaimer
Kevin suggested that i upload a sample dataset that i pulled down
to make the data acquisition process easier (ie, not require a user
to get an API key just to evaluate my code). this is a great idea.
unfortunately, twitter data does seem subject to some privacy policies,
so i do not think i can do this. i have added a disclaimer in the
readme to clarify this. please reach out to me directly for
grading or peer review purposes if this ends up proving challenging.